- Study 3 tests now work as expected - implements a separate BaseTest class for study 3 relevant tests that calls the Study 3 version of the Dialog Agent
