+++ src/main/resources/org/clulab/asist/grammars/base_concepts_study4.yml
@@ -72,7 +72,7 @@ rules:
     type: token
     keep: false
     pattern: |
-      [word=/^few$|^some$/]
+      [word=/^few$|^some$/] [word=left]?
 
   - name: coin
     priority: ${ rulepriority }
@@ -106,6 +106,14 @@ rules:
       trigger = (?<! [mention=Tool] ) [!mention=Product] [lemma=/^tool$/]  (?!  [mention=Tool])
       owner: Entity? = >"nmod:poss"
 
+  - name: tool_wire_snip
+    priority: ${ second_priority }
+    label: Tool
+    keep: ${ keep_items }
+    pattern: |
+      trigger = [word=wire] [word=/^(cutters?)|(snips?)$/]
+      owner: Entity? = >"nmod:poss"
+# TODO: figure out what they mean by wire snips
   - name: map_sectors
     priority: ${ rulepriority }
     label: MapSector
