+++ src/main/resources/org/clulab/asist/grammars/movement.yml
@@ -88,8 +88,9 @@ rules:
   - name: leave_no_location
     priority: ${ rulepriority }
     label: MoveFrom
+    keep: false
     pattern: |
-      trigger = [lemma=/(?i)^(${ movefrom_triggers })$/ & tag=/^V/ & !outgoing=/^(xcomp|dobj|nmod|ccomp)/] | [lemma=go] away
+      trigger = (?<! have ) [lemma=/(?i)^(${ movefrom_triggers })$/ & tag=/^V/ & !outgoing=/^(xcomp|dobj|nmod|ccomp)/] | [lemma=go] away
       agent: Entity? = >/${agents}/
 
   - name: enter_action
