+++ src/main/resources/org/clulab/asist/study3_study2/master.yml
@@ -54,6 +54,11 @@ rules:
 #    vars:
 #      rulepriority: "5+"
 
+  # Extract the puzzle concepts
+  - import: org/clulab/asist/study3_study2/puzzle_concepts.yml
+    vars:
+      rulepriority: "5+"
+
   # Extract time mentions, e.g., "30 seconds"
   - import: org/clulab/asist/study3_study2/time.yml
     vars:
