+++ src/main/resources/org/clulab/asist/study3_study2/team_communication.yml
@@ -152,7 +152,7 @@ rules:
     label: ToolBroken
     pattern: |
       trigger = [lemma=/(?i)broken|break/]
-      tool: Tool = >/nsubj|dobj/
+      tool: Tools = >/nsubj|dobj/
 
   - name: report_door_open
     priority: ${ rulepriority }
